# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chiptaol API Documentation",
        "description": "Open API Documentation for https://chiptaol.uz",
        "contact": {
            "email": "bakhadyrovf@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "https://chiptaol.uz"
        }
    ],
    "paths": {
...
        "/api/dashboard/cinemas": {
            "get": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Get a list of cinemas",
                "operationId": "e4720ecf51354fba325c6e5499c38b6a",
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-cinemas)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Add a new cinema",
                "operationId": "ef37ee8994342a5be4dc67d7da30b6e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CinemaStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-cinemas-specific)"
                    },
                    "422": {
                        "description": "Invalid coordinates (Unproccessable content)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Указаны неверные координаты."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/cinemas/{id}": {
            "get": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Get a specific cinema",
                "operationId": "7bdbab5f92108f33b75d255dd6ffd095",
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-cinemas-specific)"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Update an existing cinema",
                "operationId": "af8c9edc72805febb1340b2c36555c37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cinema ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CinemaStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-cinemas-specific)"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Delete an existing cinema",
                "operationId": "694698dda4361f2605eb5e0c3d609543",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cinema ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success (No Content)"
                    }
                }
            }
        },
        "/api/dashboard/cinemas/logo": {
            "post": {
                "tags": [
                    "Dashboard Cinemas"
                ],
                "summary": "Store a logo for cinema",
                "operationId": "31e209a03530fca8ec951536b101298e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "file",
                                        "example": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success (OK)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "929e45b4-084c-4160-b314-13cf7e0407d8"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/cinemas/{cinema-id}/halls": {
            "get": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Get a list of halls for specific cinema",
                "operationId": "b4c91652b5486386f67f67b81aea9326",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-halls)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Add a new hall for specific cinema",
                "operationId": "d0ca0dcf784044299be4f5baf2c4efb0",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HallStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-halls-specific)"
                    }
                }
            }
        },
        "/api/dashboard/cinemas/{cinema-id}/halls/{hall-id}": {
            "get": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Get a specific hall for specific cinema",
                "operationId": "b52145874e74e25e244acfd17d5785cd",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-halls-specific)"
                    },
                    "404": {
                        "description": "Non exists hall-id (Not Found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "No query results for model [App\\\\Models\\\\Hall] 6"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Update a specific hall for specific cinema",
                "operationId": "fcd4594e9f866b6fc55ac8f0275fa206",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HallStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-halls-specific)"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Delete a specific hall for specific cinema",
                "operationId": "e7344daaad7444bd71f65b3f8fe033b5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success (No Content)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CinemaStoreFormRequest": {
                "required": [
                    "title",
                    "address",
                    "longitude",
                    "latitude",
                    "logo_id"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Magic Cinema"
                    },
                    "address": {
                        "type": "string",
                        "example": "Some-address"
                    },
                    "logo_id": {
                        "type": "string"
                    },
                    "reference_point": {
                        "type": "string",
                        "example": "Cafe castillo olymp"
                    },
                    "longitude": {
                        "type": "numeric",
                        "example": 69.245077
                    },
                    "latitude": {
                        "type": "numeric",
                        "example": 41.326226
                    },
                    "phone": {
                        "description": "Must start with `+998`",
                        "type": "string",
                        "example": "+998909144615"
                    }
                },
                "type": "object"
            },
            "HallStoreFormRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Зал Номер - 6 (VIP)"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_vip": {
                        "type": "boolean",
                        "default": false,
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "cinema-id-path": {
                "name": "cinema-id",
                "in": "path",
                "description": "Cinema ID",
                "required": true
            },
            "hall-id-path": {
                "name": "hall-id",
                "in": "path",
                "description": "Hall ID",
                "required": true
            }
        }
    }
}